fix: minor fixes to hex view and tests

+++ src/__tests__/rv32f.test.ts
@@ -1,6 +1,6 @@
 import { rv_opcode, rv_reg, rv_reg_f } from '../hardware/riscv/riscv.const';
 import { RVProcessor } from '../hardware/riscv/riscv.processor';
-import { f_to_biguint, SP_CANONICAL_NAN, SP_NEG_INF, SP_POS_INF } from '../hardware/riscv/riscv.utils';
+import { f_to_biguint, SP_CANONICAL_NAN, SP_NEG_INF, SP_POS_INF, u32 } from '../hardware/riscv/riscv.utils';
 
 describe('RV32F', () => {
   test('flw fs10, 4(t1)', () => {
@@ -34,7 +34,7 @@ describe('RV32F', () => {
       imm: 4n,
     } as any);
 
-    expect(cpu.memoryRead(0x104n, 32)).toBe(value);
+    expect(u32(cpu.memoryRead(0x104n, 32))).toBe(value);
   });
 
   test('fmadd.s fa0, ft0, ft1, ft2', () => {
