+++ src/__tests__/rv32i.test.ts
@@ -84,7 +84,7 @@ describe('RV32I', () => {
       imm: 4n,
     } as any);
 
-    expect(cpu.memoryRead(0x104n, 32)).toBe(u32(value));
+    expect(u32(cpu.memoryRead(0x104n, 32))).toBe(u32(value));
   });
 
   describe('lb sign-extension (spec: x[rd] = sext(M[x[rs1]+offset][7:0]))', () => {
@@ -254,6 +254,6 @@ describe('RV32I', () => {
       imm: 0n,
     } as any);
 
-    expect(cpu.memoryRead(0x100n, 32)).toBe(0xdeadbeefn);
+    expect(u32(cpu.memoryRead(0x100n, 32))).toBe(0xdeadbeefn);
   });
 });
